material for session07. Here you have four files, two python files and two csv files with data. The python files load one of the csv files and uses panda and seaborn to make plots of the data.
To run the files in pycharm make sure to modify the run configuration so that it runs the file with a python console. There are two saved run configurations that I use in the .idea\runConfigurations folder of this repository, when you update or clone the repository the saved configurations should appear in the dropdown menu next to the "play" (run) icon. Choose the corresponding run configuration and you should see the plots and the variables in a python console in the bottom on the screen.